+++ exts/omni.example.defects/omni/example/defects/__init__.py
@@ -14,4 +14,3 @@
 # limitations under the License.
 
 from .extension import *
-from .ogn import *