+++ exts/omni.example.defects/docs/CHANGELOG.md
@@ -4,6 +4,32 @@ All notable changes to this project will be documented in this file.
 
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.1.0] - 2023-08-15
+
+### Removed
+
+- Ogn nodes, these nodes are now apart of the replicator pipeline
+- proxy.usd, Replicator has built in functionality in their nodes that creates the proxy
+- Functions in `utils.py` that are not longer being used
+- Region selection
+
+### Added
+
+- Options to either use bounding boxes or segmentation
+- Functionality to remove new prims created by Replicator
+- Notificataion popup for when the prim vars are applied to the mesh
+
+### Changed
+
+- Textures are now represented as a D, N, and R
+    - D is Diffuse
+    - N is Normal
+    - R is Roughness
+- Default values for dimensions start at 0.1
+- Output Directory UI defaults to replicator default output directory
+- Textures folder UI defaults to folder inside of extension with sample scratches
+- Updated README talking about the new images being used
+
 ## [1.0.1] - 2023-04-18
 
 ### Changed